% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{DFSCalculate}
\alias{DFSCalculate}
\title{DFSCalculate}
\usage{
DFSCalculate(sheet_individual, relapseDate, sorttimes, type)
}
\arguments{
\item{sheet_individual}{clinical follow up data of a individual}

\item{relapseDate}{relapse date}

\item{sorttimes}{sorted times}

\item{type}{chose FollowUp}
}
\value{
a character with disease-free survival and patient state (0: no relapse, 1: relapse)
}
\description{
DFSCalculate
}
\examples{
DFS <- DFSCalculate(sheet_individual, relapseDate, sorttimes, "FollowUp")
}
