% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{DFSProcess}
\alias{DFSProcess}
\title{DFSProcess}
\usage{
DFSProcess(sheet, ID, select, TimeInterval = 6, Ceiling = FALSE)
}
\arguments{
\item{sheet}{a combined sheet}

\item{ID}{a character of patient IDs}

\item{select}{chose FollowUp}

\item{TimeInterval}{months}

\item{Ceiling}{chose TRUE}
}
\value{
a data.frame with disease-free survival and patient state (0: no relapse, 1: relapse)
}
\description{
DFSProcess
}
\examples{
DFSAll <- DFSProcess(sheet, ID, "FollowUp", Ceiling=TRUE)
}
