% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{FollwUpRelapse}
\alias{FollwUpRelapse}
\title{FollwUpRelapse}
\usage{
FollwUpRelapse(sheet_individual)
}
\arguments{
\item{sheet_individual}{clinical follow up data of a individual}
}
\value{
a list of relapse data
}
\description{
FollwUpRelapse
}
\examples{
RawRelapseList <- FollwUpRelapse(sheet_individual)
}
