% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{OSCalculate}
\alias{OSCalculate}
\title{OSCalculate}
\usage{
OSCalculate(sheet_individual, type)
}
\arguments{
\item{sheet_individual}{clinical follow up data of a individual}

\item{type}{chose max}
}
\value{
a character with overall survival and patient state (0: survival, 1: dead)
}
\description{
OSCalculate
}
\examples{
OS <- OSCalculate(sheet_individual, "max")
}
