% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{preRelapseRecord}
\alias{preRelapseRecord}
\title{preRelapseRecord}
\usage{
preRelapseRecord(sheet_individual)
}
\arguments{
\item{sheet_individual}{clinical follow up data of a individual}
}
\value{
pre-Relapse data with record
}
\description{
preRelapseRecord
}
\examples{
preRelapseRecordDate <- preRelapseRecord(sheet_individual)
}
