% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{PreRelapseRecordDateProcess}
\alias{PreRelapseRecordDateProcess}
\title{PreRelapseRecordDateProcess}
\usage{
PreRelapseRecordDateProcess(sheet, ID)
}
\arguments{
\item{sheet}{a combined sheet}

\item{ID}{a character of patient IDs}
}
\value{
a character of pre-relapse date with record of select patients
}
\description{
PreRelapseRecordDateProcess
}
\examples{
preRelapseRecordData <- PreRelapseRecordDateProcess(sheet, ID)
}
