% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{RelapseDate}
\alias{RelapseDate}
\title{Relapse Data}
\usage{
RelapseDate(sheet_individual, type, TimeInterval)
}
\arguments{
\item{sheet_individual}{clinical follow up data of a individual}

\item{type}{chose FollowUp}

\item{TimeInterval}{months}
}
\value{
relapse data and time list
}
\description{
Relapse Data
}
\examples{
RelapseAndTimeList <- RelapseDate(sheet_individual, type, TimeInterval)
}
