% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{RelapseDateProcess}
\alias{RelapseDateProcess}
\title{RelapseDateProcess}
\usage{
RelapseDateProcess(sheet, ID, select, TimeInterval = 6)
}
\arguments{
\item{sheet}{a combined sheet}

\item{ID}{a character of patient IDs}

\item{select}{chose FollowUp}

\item{TimeInterval}{months}
}
\value{
a character of relapse date of select patients
}
\description{
RelapseDateProcess
}
\examples{
relapseDate <- RelapseDateProcess(sheet, ID, "FollowUp")
}
