% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{replication2}
\alias{replication2}
\title{replication2}
\usage{
replication2(Character)
}
\arguments{
\item{Character}{a character}
}
\value{
a list of replacated index
}
\description{
replication2
}
\examples{
ReplicateList <- replication2(Character)
}
