% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{SheetConvert}
\alias{SheetConvert}
\title{SheetConvert}
\usage{
SheetConvert(sheet0, sheet1)
}
\arguments{
\item{sheet0}{a data.frame of clinical follow up}

\item{sheet1}{a data.frame of clinical information}
}
\value{
a combined data.frame
}
\description{
SheetConvert
}
\examples{
sheet <- SheetConvert(sheet0, sheet1)
}
