% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClinFunction.R
\name{timerank}
\alias{timerank}
\title{timerank}
\usage{
timerank(sheet_individual, times)
}
\arguments{
\item{sheet_individual}{clinical follow up data of a individual}

\item{times}{time serise}
}
\value{
ranked time
}
\description{
timerank
}
\examples{
sorttimes <- timerank(sheet_individual, times)
}
